refactor(cli): physically move search handlers into cli_search.py (ARCH-2)
Real cut-and-paste of 12 search/discover handlers (_search, _websearch,
_emit_papers_input_json, _enrich, _references, _cited_by, _suggest,
_discover_*) OUT of cli.py INTO cli_search.py. cli.py 8156 -> 7686 (-470).
Shared arg parsers (_parse_year_range/_csv_terms/_negative_terms/_seed_dois)
moved to cli_common. cli.py re-imports the handlers; _sync_cli_dependencies
extended for cli_search.

Strict anti-shortcut verification (a prior 9-module batch was rejected for a
dynamic _load_cli_legacy_handlers shim with zero real defs): cli_search has 12
real def blocks, the 12 funcs are GONE from cli.py, no shim, full suite green
(3114 passed; one flaky CSRF cross-test-pollution failure cleared on re-run).
Mechanical move delegated to Codex; Claude verified + reviewed.
